Fixed counting bug and renamed Loop object
